UtilTest : Correction for linux compilation

+++ UnitTests/sources/UtilTest.cpp
@@ -66,8 +66,8 @@ ADD_TEST(distance) {
 	CHECK(Util::AddDistance(INT_MAX, 1, INT_MAX, INT_MIN) == INT_MIN);
 	CHECK(Util::AddDistance(INT_MIN, -1, INT_MAX, INT_MIN) == INT_MAX);
 	CHECK(Util::AddDistance(-95, -6, -20, -100) == -20);
-	CHECK(Util::AddDistance(LLONG_MAX, Int64(1), LLONG_MAX, Int64(1)) == 1);
-	CHECK(Util::AddDistance(UInt64(LLONG_MAX), UInt64(1), ULLONG_MAX, UInt64(1)) == UInt64(LLONG_MAX)+1);
+	CHECK(Util::AddDistance(Int64(LLONG_MAX), Int64(1), Int64(LLONG_MAX), Int64(1)) == 1);
+	CHECK(Util::AddDistance(UInt64(LLONG_MAX), UInt64(1), UInt64(ULLONG_MAX), UInt64(1)) == UInt64(LLONG_MAX)+1);
 }
